+++ winter/web/exception_handlers.py
@@ -1,5 +1,5 @@
-import typing
 from http import HTTPStatus
+from typing import Dict
 
 from winter.core.json.decoder import JSONDecodeException
 from .exceptions import ExceptionHandler
@@ -11,7 +11,7 @@ from .response_status_annotation import response_status
 
 class DecodeExceptionHandler(ExceptionHandler):
     @response_status(HTTPStatus.BAD_REQUEST)
-    def handle(self, exception: JSONDecodeException) -> typing.Dict:
+    def handle(self, exception: JSONDecodeException) -> Dict:
         return exception.errors
 
 
@@ -30,7 +30,7 @@ class BadRequestExceptionHandler(ExceptionHandler):
 
 class ThrottleExceptionHandler(ExceptionHandler):
     @response_status(HTTPStatus.TOO_MANY_REQUESTS)
-    def handle(self, exception: ThrottleException) -> typing.Dict:
+    def handle(self, exception: ThrottleException) -> Dict:
         return {
             'detail': 'Request was throttled.',
         }