+++ winter/web/urls.py
@@ -1,13 +1,13 @@
 import enum
 import inspect
 import types
-import typing
 import uuid
+from typing import Iterable
 
 import uritemplate
 
-from ..core import ComponentMethod
-from ..type_utils import get_origin_type
+from winter.core import ComponentMethod
+from winter.core.utils.typing import get_origin_type
 
 _regexp = {}
 
@@ -21,7 +21,7 @@ def register_url_regexp(func: types.FunctionType):
     return func
 
 
-def rewrite_uritemplate_with_regexps(url_path: str, methods: typing.Iterable[ComponentMethod]) -> str:
+def rewrite_uritemplate_with_regexps(url_path: str, methods: Iterable[ComponentMethod]) -> str:
     for variable_name in uritemplate.variables(url_path):
         arguments = (method.get_argument(variable_name) for method in methods)